refactor: extract collection from `client/__init__.py`
